+++ src/lib/config.ts
@@ -61,8 +61,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "0.1.0-alpha.15",
+  sdkVersion: "0.1.0-alpha.16",
   genVersion: "2.474.15",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.15 2.474.15 0.0.2 syllable-sdk",
+    "speakeasy-sdk/typescript 0.1.0-alpha.16 2.474.15 0.0.2 syllable-sdk",
 } as const;