+++ RELEASES.md
@@ -191,3 +191,13 @@ Based on:
 - [typescript v0.1.0-alpha.15] .
 ### Releases
 - [NPM v0.1.0-alpha.15] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.15 - .
+
+## 2024-12-10 19:06:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.453.9 (2.474.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.16] .
+### Releases
+- [NPM v0.1.0-alpha.16] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.16 - .